Create rental module
